rename workspace to workspace-service
