+++ src/bp/core/app.inversify.ts
@@ -16,7 +16,7 @@ import HTTPServer from './server'
 import { DataRetentionJanitor } from './services/retention/janitor'
 import { DataRetentionService } from './services/retention/service'
 import { ServicesContainerModules } from './services/services.inversify'
-import { WorkspaceService } from './services/workspace'
+import { WorkspaceService } from './services/workspace-service'
 import { Statistics } from './stats'
 import { TYPES } from './types'
