+++ src/bp/core/routers/admin/index.ts
@@ -4,8 +4,8 @@ import LicensingService from 'common/licensing-service'
 import { BotLoader } from 'core/bot-loader'
 import { GhostService } from 'core/services'
 import AuthService, { TOKEN_AUDIENCE } from 'core/services/auth/auth-service'
-import { BotService } from 'core/services/bot'
-import { WorkspaceService } from 'core/services/workspace'
+import { BotService } from 'core/services/bot-service'
+import { WorkspaceService } from 'core/services/workspace-service'
 import { RequestHandler, Router } from 'express'
 import _ from 'lodash'
