+++ src/bp/core/server.ts
@@ -23,7 +23,7 @@ import { GhostService } from './services'
 import ActionService from './services/action/action-service'
 import AuthService from './services/auth/auth-service'
 import { InvalidLicenseKey } from './services/auth/errors'
-import { BotService } from './services/bot'
+import { BotService } from './services/bot-service'
 import { CMSService } from './services/cms'
 import { ConverseService } from './services/converse'
 import { FlowService } from './services/dialog/flow/service'
@@ -31,7 +31,7 @@ import { SkillService } from './services/dialog/skill/service'
 import { LogsService } from './services/logs/service'
 import MediaService from './services/media'
 import { NotificationsService } from './services/notification/service'
-import { WorkspaceService } from './services/workspace'
+import { WorkspaceService } from './services/workspace-service'
 import { TYPES } from './types'
 
 const BASE_API_PATH = '/api/v1'
