+++ src/bp/core/services/bot-service.ts
