+++ src/bp/core/services/workspace-service.ts
